#242 Migrated to .NET Standard. Removed unused JsonObject.cs file, added solution file and moved project files to subfolder
