--- WooCommerceNET.csproj
@@ -1,103 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <MinimumVisualStudioVersion>10.0</MinimumVisualStudioVersion>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{63DDA1FE-3011-4700-A0F4-78E2358B330F}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>WooCommerceNET</RootNamespace>
-    <AssemblyName>WooCommerceNET</AssemblyName>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <FileAlignment>512</FileAlignment>
-    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <TargetFrameworkProfile>Profile259</TargetFrameworkProfile>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <NuGetPackageImportStamp>
-    </NuGetPackageImportStamp>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <!-- A reference to the entire .NET Framework is automatically included -->
-    <None Include="Changes.md" />
-    <None Include="License.md" />
-    <None Include="Readme.md" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Base\Common.cs" />
-    <Compile Include="Base\Extension.cs" />
-    <Compile Include="Base\HMAC-SHA256.cs" />
-    <Compile Include="Base\BaseObject.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RestAPI.cs" />
-    <Compile Include="Base\SHA1.cs" />
-    <Compile Include="WooCommerce\v1\Address.cs" />
-    <Compile Include="WooCommerce\v1\Coupon.cs" />
-    <Compile Include="WooCommerce\v1\Customer.cs" />
-    <Compile Include="WooCommerce\Legacy\Coupon.cs" />
-    <Compile Include="WooCommerce\Legacy\Customer.cs" />
-    <Compile Include="WooCommerce\Legacy\Order.cs" />
-    <Compile Include="WooCommerce\Legacy\Product.cs" />
-    <Compile Include="WooCommerce\Legacy\Store.cs" />
-    <Compile Include="WooCommerce\Legacy\Tax.cs" />
-    <Compile Include="WooCommerce\Legacy\WCObject.cs" />
-    <Compile Include="WooCommerce\Legacy\Webhook.cs" />
-    <Compile Include="WooCommerce\v1\Order.cs" />
-    <Compile Include="WooCommerce\v1\Product.cs" />
-    <Compile Include="WooCommerce\v1\Report.cs" />
-    <Compile Include="WooCommerce\v1\Tax.cs" />
-    <Compile Include="WooCommerce\v1\WCObject.cs" />
-    <Compile Include="WooCommerce\v1\Webhook.cs" />
-    <Compile Include="WooCommerce\v2\ProductAttribute.cs" />
-    <Compile Include="WooCommerce\v2\ProductCategory.cs" />
-    <Compile Include="WooCommerce\v2\Customer.cs" />
-    <Compile Include="WooCommerce\v2\Coupon.cs" />
-    <Compile Include="WooCommerce\v2\Order.cs" />
-    <Compile Include="WooCommerce\v2\OrderNote.cs" />
-    <Compile Include="WooCommerce\v2\PaymentGateway.cs" />
-    <Compile Include="WooCommerce\v2\Product.cs" />
-    <Compile Include="WooCommerce\v2\OrderRefund.cs" />
-    <Compile Include="WooCommerce\v2\Report.cs" />
-    <Compile Include="WooCommerce\v2\Setting.cs" />
-    <Compile Include="WooCommerce\v2\ShippingClass.cs" />
-    <Compile Include="WooCommerce\v2\ShippingMethod.cs" />
-    <Compile Include="WooCommerce\v2\ShippingZone.cs" />
-    <Compile Include="WooCommerce\v2\SystemStatus.cs" />
-    <Compile Include="WooCommerce\v2\ProductTag.cs" />
-    <Compile Include="WooCommerce\v2\TaxClass.cs" />
-    <Compile Include="WooCommerce\v2\TaxRate.cs" />
-    <Compile Include="WooCommerce\v2\Variation.cs" />
-    <Compile Include="WooCommerce\v2\WCObject.cs" />
-    <Compile Include="WooCommerce\v2\Webhook.cs" />
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
-  <PropertyGroup>
-    <PostBuildEvent>
-    </PostBuildEvent>
-  </PropertyGroup>
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
