+++ WooCommerceNET.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2024
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WooCommerceNET", "WooCommerceNET\WooCommerceNET.csproj", "{99144D39-9137-499C-BBE3-A504112BDD6F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{99144D39-9137-499C-BBE3-A504112BDD6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99144D39-9137-499C-BBE3-A504112BDD6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{99144D39-9137-499C-BBE3-A504112BDD6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{99144D39-9137-499C-BBE3-A504112BDD6F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E7E67CCD-5E9B-48FE-BFE6-5830D2DA9DD5}
+	EndGlobalSection
+EndGlobal
