+++ WooCommerceNET/Base/BaseObject.cs
