+++ WooCommerceNET/Base/Common.cs
