+++ WooCommerceNET/Base/Extension.cs
