+++ WooCommerceNET/Base/HMAC-SHA256.cs
