+++ WooCommerceNET/Base/SHA1.cs
