+++ WooCommerceNET/WooCommerce/Legacy/Coupon.cs
