+++ WooCommerceNET/WooCommerce/Legacy/Product.cs
