+++ WooCommerceNET/WooCommerce/Legacy/Store.cs
