+++ WooCommerceNET/WooCommerce/Legacy/Tax.cs
