+++ WooCommerceNET/WooCommerce/Legacy/WCObject.cs
