+++ WooCommerceNET/WooCommerce/v2/Product.cs
