+++ WooCommerceNET/WooCommerce/v2/ProductCategory.cs
