+++ WooCommerceNET/WooCommerce/v2/Setting.cs
