+++ WooCommerceNET/WooCommerce/v2/ShippingClass.cs
