+++ WooCommerceNET/WooCommerce/v2/TaxClass.cs
