+++ WooCommerceNET/WooCommerce/v2/Variation.cs
